Update resubmission function for new submission flow
