+++ questionnaire/questionnaire-service.test.js
@@ -945,81 +945,5 @@ describe('Questionnaire Service', () => {
                 );
             });
         });
-
-        describe('postFailedSubmissions', () => {
-            it('Should resubmit a failed application', async () => {
-                jest.doMock('./questionnaire-dal', () => {
-                    const dalServiceMock = {
-                        getQuestionnaireIdsBySubmissionStatus: jest.fn(() => {
-                            return [validQuestionnaireId];
-                        }),
-                        updateQuestionnaireSubmissionStatus: jest.fn(() => {
-                            return 'ok!';
-                        })
-                    };
-
-                    return () => dalServiceMock;
-                });
-                const createQuestionnaireService = require('./questionnaire-service');
-                const questionnaireService = createQuestionnaireService({
-                    logger: {info: jest.fn()},
-                    apiVersion: undefined, // Undefined should only occur for DCS API v1 and Admin.
-                    ownerId: undefined // Undefined should only occur for DCS API v1 and Admin.
-                });
-                const actual = await questionnaireService.postFailedSubmissions();
-
-                expect(actual).toMatchObject([
-                    {
-                        id: validQuestionnaireId,
-                        resubmitted: true
-                    }
-                ]);
-            });
-
-            it('Should do nothing if no failed submissions are found', async () => {
-                jest.doMock('./questionnaire-dal', () => {
-                    const dalServiceMock = {
-                        getQuestionnaireIdsBySubmissionStatus: jest.fn(() => {
-                            return [];
-                        }),
-                        updateQuestionnaireSubmissionStatus: jest.fn(() => {
-                            return 'ok!';
-                        })
-                    };
-
-                    return () => dalServiceMock;
-                });
-                const createQuestionnaireService = require('./questionnaire-service');
-                const questionnaireService = createQuestionnaireService({
-                    logger: {info: jest.fn()},
-                    apiVersion: undefined, // Undefined should only occur for DCS API v1 and Admin.
-                    ownerId: undefined // Undefined should only occur for DCS API v1 and Admin.
-                });
-                const actual = await questionnaireService.postFailedSubmissions();
-                expect(actual).toMatchObject([]);
-            });
-
-            it('Should error gracefully', async () => {
-                jest.doMock('./questionnaire-dal', () => {
-                    const dalServiceMock = {
-                        getQuestionnaireIdsBySubmissionStatus: jest.fn(() => {
-                            return undefined;
-                        }),
-                        updateQuestionnaireSubmissionStatus: jest.fn(() => {
-                            return 'ok!';
-                        })
-                    };
-
-                    return () => dalServiceMock;
-                });
-                const createQuestionnaireService = require('./questionnaire-service');
-                const questionnaireService = createQuestionnaireService({
-                    logger: {info: jest.fn(), error: jest.fn()},
-                    apiVersion: undefined, // Undefined should only occur for DCS API v1 and Admin.
-                    ownerId: undefined // Undefined should only occur for DCS API v1 and Admin.
-                });
-                await expect(questionnaireService.postFailedSubmissions()).rejects.toThrow();
-            });
-        });
     });
 });
